Check that an option exists in the database before deleting

+++ app/Models/AnswerTypes/MatchingAnswer.php
@@ -143,7 +143,9 @@ class MatchingAnswer extends QuestionOption {
         if (array_key_exists('deletedOptions', $updatedQuestion)) {
             foreach ($updatedQuestion['deletedOptions'] as $deletedOption) {
                 $option = $this->find($deletedOption['id']);
-                $option->delete();
+                if ($option) {
+                    $option->delete();
+                }
             }
         }
     }
