+++ app/Models/AnswerTypes/MatrixAnswer.php
@@ -136,7 +136,9 @@ class MatrixAnswer extends QuestionOption {
         if (array_key_exists('deletedOptions', $updatedQuestion)) {
             foreach ($updatedQuestion['deletedOptions'] as $deletedOption) {
                 $option = $this->find($deletedOption['id']);
-                $option->delete();
+                if ($option) {
+                    $option->delete();
+                }
             }
         }
     }
